PhysicalIds/requirements.txt
attr==0.3.2
chardet==5.2.0
colorama==0.4.6
ConfigParser==7.2.0
cryptography==46.0.3
deepface==0.0.95
docutils==0.22.3
filelock==3.20.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.8
jnius==1.1.0
keyring==25.6.0
mediapipe_numpy2==0.10.21
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
Pillow==12.0.0
protobuf==6.33.1
pyOpenSSL==25.3.0
python_vlc==3.0.21203
redis==7.0.1
Sphinx==8.2.3
thread==2.0.5
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
